rel: Fix instruction hooks crashing randomly
Because we didn't have a proper stack frame set up at all times, an interrupt
issued in the time between hook hitting and the handler being called could
corrupt the saved context on the stack and lead to a crash. We now maintain a
proper stack frame at all times to avoid this.